Move source to root
