--- CTrL/docs/source/_autoapi_templates/index.rst
@@ -1,17 +1,0 @@
-API Reference
-=============
-
-CTrL
-----
-
-.. toctree::
-   :titlesonly:
-   :maxdepth: 2
-
-
-{% for page in pages %}
-{% if page.top_level_object and page.display %}
-{{ page.include_path }}
-{% endif %}
-{% endfor %}
-
