--- CTrL/docs/source/conf.py
@@ -1,85 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates.
-# 
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-import os
-import sys
-# sys.path.insert(0, os.path.abspath('..'))
-sys.path.insert(0, os.path.abspath('../..'))
-
-
-# -- Project information -----------------------------------------------------
-
-project = 'CTrL Benchmark'
-copyright = "2020, Facebook, Inc. and its affiliates"
-author = "Tom Veniat, Ludovic Denoyer & Marc'Aurelio Ranzato"
-
-# The full version, including alpha/beta/rc tags
-release = '0.0.2'
-
-
-# -- General configuration ---------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-
-# extensions = [
-#     'sphinx.ext.autodoc',  # Core library for html generation from docstrings
-#     'sphinx.ext.autosummary',  # Create neat summary tables
-# ]
-# autosummary_generate = True  # Turn on sphinx.ext.autosummary
-
-extensions = [
-    # 'sphinx.ext.autodoc',
-    # 'sphinx.ext.autodoc.typehints',
-    'autoapi.extension',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.autodoc',
-    'sphinx_autodoc_typehints'
-]
-
-autoapi_type = 'python'
-autoapi_dirs = ['../../ctrl']
-
-# autoapi_template_dir = '_autoapi_templates'
-autoapi_python_class_content = 'both'
-autodoc_typehints = 'description'
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
-
-
-# -- Options for HTML output -------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-# html_theme = 'alabaster'
-# html_theme = 'default'
-on_rtd = os.environ.get("READTHEDOCS", None) == "True"
-if not on_rtd:  # only import and set the theme if we're building docs locally
-    import sphinx_rtd_theme
-    html_theme = "sphinx_rtd_theme"
-    html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
