--- CTrL/docs/source/index.rst
@@ -1,26 +1,0 @@
-.. CTrL Benchmark documentation master file, created by
-   sphinx-quickstart on Thu Oct 22 18:00:13 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-   https://stackoverflow.com/a/62613202/5629805
-
-Welcome to CTrL Benchmark's documentation!
-==========================================
-
-The Continual Transfer Learning Benchmark is a framework agnostic project
-aiming to facilitate the evaluation of the different properties wanted for a
-lifelong learning model.
-
-.. toctree::
-   :titlesonly:
-   :maxdepth: 4
-
-   usage
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
