--- CTrL/docs/source/usage.rst
@@ -1,9 +1,0 @@
-Usage
-=====
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   usage/installation
-   usage/quickstart
