--- CTrL/docs/source/usage/installation.rst
@@ -1,10 +1,0 @@
-Installation
-============
-
-The CTrL bencmark can be easily installed using pip::
-
-    pip install git+git://github.com/TomVeniat/ctrl.git
-
-or when the package will be available on Pypi::
-
-    pip install ctrl-benchmark
