--- CTrL/setup.py
@@ -1,13 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates.
-# 
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-import setuptools
-
-setuptools.setup(
-    name="ctrl-benchmark",
-    version="0.0.2",
-    author="Tom Veniat, Ludovic Denoyer & Marc'Aurelio Ranzato",
-    description="Continual Transfer Learning Benchmark",
-    packages=setuptools.find_packages(),
-)
