+++ ctrl/instances/mnist.py
