+++ ctrl/instances/synthetic_gaussians.py
